+++ threedigrid/admin/serializers.py
@@ -142,7 +142,7 @@ def fill_properties(fields, data, index, model_type=None):
     :param model_type: (str, optional) optionally adds the given model_type
     """
     result = OrderedDict()
-    for i, field in enumerate(fields):
+    for field in fields:
         if isinstance(field, dict):
             for key, sub_list in field.items():
                 result[key] = fill_properties(sub_list, data, index)
@@ -150,16 +150,9 @@ def fill_properties(fields, data, index, model_type=None):
             field_data = data.get(field, None)
             if field_data is not None and field_data.size > 0:
                 value = field_data[..., index]
-                # Replace 'no data' values with None (NULL) for all dtypes:
-                # - floats: NaN, Inf, -Inf, -9999.0
-                # - ints: -9999
-                # - str/bytes: ""
+                # Replace NaN, Inf, -Inf, -9999.0 floats with None (becomes JSON null)
                 if np.issubdtype(value.dtype, np.floating):
                     is_invalid = (~np.isfinite(value)) | (value == -9999.0)
-                elif np.issubdtype(value.dtype, np.integer):
-                    is_invalid = value == -9999
-                elif np.issubdtype(value.dtype, np.character):
-                    is_invalid = np.char.str_len(value) == 0
                 else:
                     is_invalid = False
                 if np.any(is_invalid):